+++ lib/dns/sec/openssl/a_int.c
@@ -340,7 +340,7 @@ int ASN1_INTEGER_set(ASN1_INTEGER *a, long v)
 	long d;
 
 	a->type=V_ASN1_INTEGER;
-	if (a->length < (sizeof(long)+1))
+	if (a->length < (int)(sizeof(long)+1))
 		{
 		if (a->data != NULL)
 			Free(a->data);
@@ -359,7 +359,7 @@ int ASN1_INTEGER_set(ASN1_INTEGER *a, long v)
 		a->type=V_ASN1_NEG_INTEGER;
 		}
 
-	for (i=0; i<sizeof(long); i++)
+	for (i=0; i<(int)sizeof(long); i++)
 		{
 		if (d == 0) break;
 		buf[i]=(int)d&0xff;
@@ -384,7 +384,7 @@ long ASN1_INTEGER_get(ASN1_INTEGER *a)
 	else if (i != V_ASN1_INTEGER)
 		return(0);
 	
-	if (a->length > sizeof(long))
+	if (a->length > (int)sizeof(long))
 		{
 		/* hmm... a bit ugly */
 		return(0xffffffffL);
